+++ next.config.ts
@@ -15,15 +15,6 @@ const nextConfig: NextConfig = {
       },
     ],
   },
-  async redirects() {
-    return [
-      {
-        source: '/en/:path*',
-        destination: '/:path*',
-        permanent: true,
-      },
-    ]
-  },
 }
 
 export default withNextIntl(nextConfig)